refactor: validate 통합
