+++ src/utils/validations/nameValidator.ts
@@ -1,5 +1,5 @@
+import { PW_DIGIT } from '@src/constants';
 import { ERROR_MESSAGE } from '@src/constants/message';
-import type { Validator } from './types';
 
 function hasBlank(str: string) {
   return str.trim() !== str;
@@ -9,22 +9,43 @@ function isEmpty(str: string) {
   return str === '';
 }
 
+function isNotProperDigit(num: number) {
+  return num.toString().length !== PW_DIGIT;
+}
+
+interface Validator<T> {
+  test: (str: T) => boolean;
+  message: string;
+}
+
+type Validators = { [P in 'name' | 'pw']: (Validator<number> | Validator<string>)[] };
+
 // @TODO: 이름 중복 체크 추가
-const validators: Validator<string>[] = [
-  {
-    test: hasBlank,
-    message: ERROR_MESSAGE.HAS_BLANK,
-  },
-  {
-    test: isEmpty,
-    message: ERROR_MESSAGE.EMPTY,
-  },
-];
+const validators: Validators = {
+  name: [
+    {
+      test: isEmpty,
+      message: ERROR_MESSAGE.EMPTY,
+    },
+    {
+      test: hasBlank,
+      message: ERROR_MESSAGE.HAS_BLANK,
+    },
+  ],
+  pw: [
+    {
+      test: isNotProperDigit,
+      message: ERROR_MESSAGE.NOT_PROPER_DIGIT,
+    },
+  ],
+};
 
 // @TODO: never | void type 지정
-export function nameValidator(name: string) {
-  return validators.every(({ test, message }) => {
-    if (test(name)) {
+export function validate(type: keyof Validators, value: any) {
+  validators[type].every;
+
+  return validators[type].every(({ test, message }) => {
+    if (test(value as never)) {
       throw new Error(message);
     }
     return true;
