--- src/utils/validations/types.ts
@@ -1,6 +1,0 @@
-interface Validator<T> {
-  test: (str: T) => boolean;
-  message: string;
-}
-
-export type { Validator };
